Moved the "preventLeavingDialog" function from "DialogUtil" to "HtmlUtil" for better organization. Extracted a new constant "INTERACTIVE_ELEMENTS_SELECTOR", which is selector for all interactive elements.
